+++ starboard/shared/starboard/command_line_preprocessor.h
@@ -1,0 +1,29 @@
+// Copyright 2026 The Cobalt Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef STARBOARD_SHARED_STARBOARD_COMMAND_LINE_PREPROCESSOR_H_
+#define STARBOARD_SHARED_STARBOARD_COMMAND_LINE_PREPROCESSOR_H_
+
+#include "starboard/common/command_line.h"
+
+namespace starboard {
+
+// Preprocesses command-line switches by converting deprecated legacy flags
+// (e.g., --dump_video_data) into equivalent --enable-features and
+// --disable-features switches.
+void ConvertDeprecatedSwitches(CommandLine* command_line);
+
+}  // namespace starboard
+
+#endif  // STARBOARD_SHARED_STARBOARD_COMMAND_LINE_PREPROCESSOR_H_
